requirements.txt
scipy
scikit-learn
scikit-image
numpy
opencv_python
Pillow
torch>=1.2.0
torchvision
tensorboardX
tqdm
regex
imageio